fix bad imports in extension-shared.ts
