+++ src/extension/constants.ts
@@ -1,0 +1,14 @@
+/*
+ * constants.ts
+ *
+ * Copyright (C) 2020-2022 Posit Software, PBC
+ */
+
+export const kBuiltInExtOrg = "quarto";
+export const kCommon = "common";
+export const kExtensionDir = "_extensions";
+export const kTitle = "title";
+export const kAuthor = "author";
+export const kVersion = "version";
+export const kQuartoRequired = "quarto-required";
+export const kRevealJSPlugins = "revealjs-plugins";
