+++ src/quarto-core/inspect.ts
@@ -1,9 +1,8 @@
 /*
-* inspect.ts
-*
-* Copyright (C) 2020-2022 Posit Software, PBC
-*
-*/
+ * inspect.ts
+ *
+ * Copyright (C) 2020-2022 Posit Software, PBC
+ */
 
 import { existsSync } from "fs/mod.ts";
 import { dirname, join, relative } from "path/mod.ts";
@@ -27,7 +26,7 @@ import { ProjectConfig, ProjectFiles } from "../project/types.ts";
 import { cssFileResourceReferences } from "../core/css.ts";
 import { projectExcludeDirs } from "../project/project-shared.ts";
 import { normalizePath, safeExistsSync } from "../core/path.ts";
-import { kExtensionDir } from "../extension/extension-shared.ts";
+import { kExtensionDir } from "../extension/constants.ts";
 import { extensionFilesFromDirs } from "../extension/extension.ts";
 
 export interface InspectedConfig {
