+++ src/resources/editor/tools/vs-code.mjs
@@ -21059,8 +21059,7 @@ var require_yaml_intelligence_resources = __commonJS({
           long: "Title of the volume of the item or container holding the item.\nAlso use for titles of periodical special issues, special sections,\nand the like."
         },
         "Disambiguating year suffix in author-date styles (e.g.&nbsp;\u201Ca\u201D in \u201CDoe,\n1999a\u201D).",
-        "internal-schema-hack",
-        "Specifies the depth of items in the table of contents that should be\ndisplayed as expanded in HTML output. Use <code>true</code> to expand\nall or <code>false</code> to collapse all."
+        "internal-schema-hack"
       ],
       "schema/external-schemas.yml": [
         {
@@ -21284,12 +21283,12 @@ var require_yaml_intelligence_resources = __commonJS({
         mermaid: "%%"
       },
       "handlers/mermaid/schema.yml": {
-        _internalId: 152478,
+        _internalId: 148648,
         type: "object",
         description: "be an object",
         properties: {
           "mermaid-format": {
-            _internalId: 152470,
+            _internalId: 148640,
             type: "enum",
             enum: [
               "png",
@@ -21305,7 +21304,7 @@ var require_yaml_intelligence_resources = __commonJS({
             exhaustiveCompletions: true
           },
           theme: {
-            _internalId: 152477,
+            _internalId: 148647,
             type: "anyOf",
             anyOf: [
               {
