+++ src/resources/editor/tools/yaml/yaml-intelligence-resources.json
@@ -14035,8 +14035,7 @@
       "long": "Title of the volume of the item or container holding the item.\nAlso use for titles of periodical special issues, special sections,\nand the like."
     },
     "Disambiguating year suffix in author-date styles (e.g.&nbsp;“a” in “Doe,\n1999a”).",
-    "internal-schema-hack",
-    "Specifies the depth of items in the table of contents that should be\ndisplayed as expanded in HTML output. Use <code>true</code> to expand\nall or <code>false</code> to collapse all."
+    "internal-schema-hack"
   ],
   "schema/external-schemas.yml": [
     {
@@ -14260,12 +14259,12 @@
     "mermaid": "%%"
   },
   "handlers/mermaid/schema.yml": {
-    "_internalId": 152478,
+    "_internalId": 148648,
     "type": "object",
     "description": "be an object",
     "properties": {
       "mermaid-format": {
-        "_internalId": 152470,
+        "_internalId": 148640,
         "type": "enum",
         "enum": [
           "png",
@@ -14281,7 +14280,7 @@
         "exhaustiveCompletions": true
       },
       "theme": {
-        "_internalId": 152477,
+        "_internalId": 148647,
         "type": "anyOf",
         "anyOf": [
           {
